# docs/swagger.yaml
definitions:
  app.Display:
    properties:
      color:
        description: 'pilot light colour when on: green, red or amber'
        type: string
      label:
        description: the relay name unless configured
        type: string
      offText:
        description: word for the off state, OFF when empty
        type: string
      onText:
        description: word for the on state, ON when empty
        type: string
    type: object
  app.HTTPResponse:
    properties:
      description:
        type: string
      display:
        $ref: '#/definitions/app.Display'
      gpio:
        type: integer
      lastChange:
        allOf:
        - $ref: '#/definitions/app.LastChangeInfo'
        description: absent when unknown
      lock:
        allOf:
        - $ref: '#/definitions/app.LockInfo'
        description: absent without minSwitchInterval
      name:
        type: string
      state:
        type: string
    type: object
  app.LastChangeInfo:
    properties:
      ageSeconds:
        description: seconds since Time, by the Pi's clock
        type: number
      client:
        description: IP address of the API client
        type: string
      host:
        description: reverse DNS name of the client, when it has one
        type: string
      source:
        description: api, or start when relayhat switched it on start
        type: string
      time:
        description: RFC 3339, in the Pi's time zone
        type: string
    type: object
  app.LockInfo:
    properties:
      intervalSeconds:
        description: minSwitchInterval
        type: number
      remainingSeconds:
        description: until it may be switched again, 0 when it may
        type: number
    type: object
  health.Model:
    properties:
      app:
        description: Application name (MODULE)
        type: string
      appVersion:
        description: Current version of the application
        type: string
      goVersion:
        description: Go runtime version
        type: string
      heapAllocBytes:
        description: Allocated heap memory in bytes
        type: integer
      hostname:
        description: Machine name where the app runs
        type: string
      numGoroutines:
        description: Current number of active goroutines
        type: integer
      os:
        description: Operating system name
        type: string
      sysMemoryBytes:
        description: Total memory obtained from the OS
        type: integer
      timestamp:
        description: UTC timestamp when health info was collected (RFC3339)
        type: string
      uptimeSeconds:
        description: Application uptime in seconds
        type: number
    type: object
  web.ApiError:
    properties:
      error:
        type: string
    type: object
info:
  contact: {}
paths:
  /health:
    get:
      description: Retrieves memory usage, goroutine count, version, hostname, Go
        runtime version, and OS.
      produces:
      - application/json
      responses:
        "200":
          description: Health data successfully retrieved
          schema:
            $ref: '#/definitions/health.Model'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Get health data
      tags:
      - info
  /relays:
    get:
      description: Returns name, state (on/off) and description for every configured
        relay.
      produces:
      - application/json
      responses:
        "200":
          description: List of all relays
          schema:
            items:
              $ref: '#/definitions/app.HTTPResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ApiError'
      security:
      - ApiKeyAuth: []
      summary: List all relays
      tags:
      - relay
  /relays/{name}:
    get:
      description: Returns name, state (on/off) and description for the given relay.
      parameters:
      - description: Relay name (e.g. relay1)
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Relay state
          schema:
            $ref: '#/definitions/app.HTTPResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ApiError'
        "404":
          description: Relay not found
          schema:
            $ref: '#/definitions/web.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Get relay state
      tags:
      - relay
  /relays/{name}/{state}:
    patch:
      description: Sets the state of a specific relay by name.
      parameters:
      - description: Relay name
        in: path
        name: name
        required: true
        type: string
      - description: Relay state (on/off)
        in: path
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Relay state successfully set
          schema:
            $ref: '#/definitions/app.HTTPResponse'
        "400":
          description: Invalid state
          schema:
            $ref: '#/definitions/web.ApiError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ApiError'
        "404":
          description: Relay not found
          schema:
            $ref: '#/definitions/web.ApiError'
        "429":
          description: Switching locked by minSwitchInterval, see the Retry-After
            header
          schema:
            $ref: '#/definitions/web.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ApiError'
        "503":
          description: Shutting down or reloading, try again
          schema:
            $ref: '#/definitions/web.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Set relay state
      tags:
      - relay
  /version:
    get:
      description: Returns the current application name and version. No authentication
        required.
      produces:
      - application/json
      responses:
        "200":
          description: Application version and name
          schema:
            properties:
              app:
                type: string
              appVersion:
                type: string
            type: object
      summary: Get application version and name
      tags:
      - info
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"